group curves by cage and correct a minor bug / csv
when computing kymos, it asked to save data in a non-existent directory

+++ src/main/java/plugins/fmp/multicafe2/experiment/Capillaries.java
@@ -6,6 +6,7 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -686,6 +687,10 @@ public class Capillaries
 	
 	private boolean csvSaveCapillariesMeasures_Data(String directory) 
 	{
+		Path path = Paths.get(directory);
+		if (!Files.exists(path))
+			return false;
+		
 		try {
 			FileWriter csvWriter = new FileWriter(directory + File.separator +"CapillariesMeasures.csv");
